feat(33-05): x64 link blocks for gl05/06/07 D3D9 plugins + DllExport x64 port
- Direct3d9/_ffp/_vsps: Debug|x64 + Release|x64 ItemDefinitionGroups (import
  x64-platform.props, isolated x64 OutDir, FirstDirect3d9 PCH Create on x64),
  x64 Link blocks: d3dx9.lib DROPPED, MachineX64, libjpeg->jpeg62-x64.lib,
  d3d9/d3dcompiler/ddraw from the Windows SDK x64, no x86 BaseAddress,
  PostBuild->stage-x64\ + the x64 System32 d3dcompiler_47.dll.
- DllExport (the plugin<->host engine-symbol bridge): ported x86 __asm int 3
  export stubs to a portable DLLEXPORT_UNREACHABLE_TRAP (__debugbreak on x64);
  added Debug|x64 + Release|x64 DynamicLibrary configs so the plugins resolve
  their engine dllimport refs against DllExport.lib (Rule 3 blocking dep).
- BITS-02 x64 width fixes in the plugin source (Rule 1): FIELD_OFFSET cast
  through size_t (Direct3d9_LightManager.cpp); pBits pointer arithmetic via
  size_t (Direct3d9.cpp); dxerr9 (no x64 lib) gated out, x64 HRESULT fallback
  for FATAL_DX_HR (Direct3d9.h).
- All 3 plugins link x64: 0 unresolved external symbol, machine 8664; Win32
  blocks additive-only (non-regression).

+++ src/engine/client/application/Direct3d9/src/win32/Direct3d9.cpp
@@ -2750,7 +2750,7 @@ bool Direct3d9Namespace::applyGammaCorrectionToXRGBSurface( IDirect3DSurface9 *s
 	for( unsigned nLine = 0; nLine != desc.Height; nLine++ )
 	{
 		// get the start of the line
-		PackedArgb * pBuffer = (PackedArgb *)((unsigned int)lockedRect.pBits + lockedRect.Pitch * nLine);
+		PackedArgb * pBuffer = (PackedArgb *)((size_t)lockedRect.pBits + lockedRect.Pitch * nLine);   // Phase 33 (BITS-02): pointer-width arithmetic (was (unsigned int), x64-truncating under /we4311/4312)
 
 		// color correct the line
 		PackedArgb * pBufferEol = pBuffer + desc.Width;
